Fixed tranlsations
