+++ src/components/app-card/AppCard.styles.js
@@ -1,0 +1,13 @@
+import { commonHoverShadow, commonShadow } from '~/styles/app-theme/custom-shadows'
+
+export const styles = {
+  container: (isClickable) => ({
+    display: 'flex',
+    padding: '30px 20px',
+    boxShadow: commonShadow,
+    borderRadius: '6px',
+    '&:hover': isClickable && {
+      boxShadow: commonHoverShadow
+    }
+  })
+}
