+++ src/components/app-chip/AppChip.tsx
@@ -7,15 +7,14 @@ import { SxProps } from '@mui/system'
 import { styles } from '~/components/app-chips-list/AppChipsList-styles'
 
 interface AppChipProps {
-  handleDelete:() => void,
-  children:ReactNode,
-  icon:React.ReactElement<SvgIconProps>,
+  handleDelete: () => void,
+  children: ReactNode,
+  icon: React.ReactElement<SvgIconProps>,
   sx?: SxProps,
+  labelSx?: SxProps
 }
 
-
-const AppChip:React.FC<AppChipProps> = ({ handleDelete, children, icon, sx }) => {
-  
+const AppChip: React.FC<AppChipProps> = ({ handleDelete, children, icon, sx, labelSx }) => {
   return (
     <Chip
       data-testid='chip'
@@ -27,7 +26,7 @@ const AppChip:React.FC<AppChipProps> = ({ handleDelete, children, icon, sx }) =>
         </IconButton>
       }
       icon={ icon }
-      label={ <Typography sx={ { typography: 'subtitle2' } }>
+      label={ <Typography sx={ { typography: 'subtitle2', ...labelSx } }>
         { children }
       </Typography> }
       onDelete={ handleDelete }
