+++ src/components/offer-card/offer-actions/OfferActions.styles.js
@@ -1,0 +1,27 @@
+export const styles = {
+  container: {
+    minWidth: '166px'
+  },
+  containerTop: {
+    position: 'relative',
+    mb: '30px'
+  },
+  bookmarkButton: {
+    position: 'absolute',
+    right: 0,
+    top: 0,
+    color: 'basic.black'
+  },
+  sendMessageButton: {
+    width: '100%',
+    mb: '16px'
+  },
+  viewDetailsButton: {
+    width: '100%',
+    backgroundColor: 'primary.50',
+    color: 'primary.900',
+    '&:hover': {
+      backgroundColor: 'primary.50'
+    }
+  }
+}
