+++ src/components/offer-card/offer-avatar-and-rating/OfferAvatarAndRating.styles.js
@@ -1,0 +1,16 @@
+export const styles = {
+  container: {
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    mr: '20px'
+  },
+  avatar: {
+    height: '80px',
+    width: '80px',
+    mb: '16px'
+  },
+  rating: {
+    backgroundColor: 'primary.50'
+  }
+}
