+++ src/components/offer-card/offer-details/OfferDetails.styles.js
@@ -1,0 +1,54 @@
+const ellipsisTextStyle = (linesCount) => ({
+  display: '-webkit-box',
+  WebkitLineClamp: linesCount,
+  lineClamp: linesCount,
+  WebkitBoxOrient: 'vertical',
+  boxOrient: 'vertical',
+  overflow: 'hidden'
+})
+
+export const styles = {
+  container: {
+    mr: '50px'
+  },
+  bio: {
+    ...ellipsisTextStyle(2),
+    mb: '10px'
+  },
+  chipsContainer: {
+    display: 'flex',
+    mb: '10px'
+  },
+  subjectChip: {
+    mr: '4px',
+    backgroundColor: 'green.300',
+    color: 'green.900'
+  },
+  levelChip: {
+    backgroundColor: 'green.50',
+    color: 'primary.700'
+  },
+  subjectChipLabel: {
+    typography: 'overline',
+    fontWeight: 500
+  },
+  levelChipLabel: {
+    typography: 'overline'
+  },
+  description: {
+    ...ellipsisTextStyle(4),
+    color: 'primary.600',
+    mb: '10px'
+  },
+  languagesContainer: {
+    display: 'flex',
+    alignItems: 'center'
+  },
+  languageIcon: {
+    color: 'primary.400',
+    marginRight: '8px'
+  },
+  languages: {
+    color: 'primary.400'
+  }
+}
