--- src/pages/find-tutor/FindTutor.jsx
@@ -1,22 +1,0 @@
-import OfferCard from '~/components/offer-card/OfferCard'
-
-const mockOffer = {
-  imgSrc: 'https://images.unsplash.com/photo-1570295999919-56ceb5ecca61?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=880&q=80',
-  rating: 4.3,
-  name: 'Vasya Pupkin',
-  bio: 'Senior lecturer at the Department of German Philology and Translation Department of English Philology Senior lecturer at the Department of German Philology and Translation Department of English Philology Senior lecturer at the Department of German Philology and Translation Department of English Philology',
-  description: 'Hello. There are many variations of passages of Lorem Ipsum available, but the majority have suffered alteration in some form, by injected humour, or randomised words which.',
-  languages: ['Ukrainian', 'English'],
-  price: 100,
-  isBookmarked: false,
-  subject: 'English',
-  level: 'Advanced'
-}
-
-const FindTutor = () => {
-  return (
-    <OfferCard {...mockOffer} />
-  )
-}
-
-export default FindTutor
